+++ Sources/VComponents/Helpers/Logging.swift
@@ -10,7 +10,6 @@ import OSLog
 
 // MARK: - Loggers
 extension Logger {
-    static let bottomSheet: Self = .init(subsystem: "VComponents", category: "VBottomSheet")
     static let compactPageIndicator: Self = .init(subsystem: "VComponents", category: "VCompactPageIndicator")
     static let rangeSlider: Self = .init(subsystem: "VComponents", category: "VRangeSlider")
     static let rollingCounter: Self = .init(subsystem: "VComponents", category: "VRollingCounter")